Refactoring things over to uk.codecymru.drawbox namespace
