--- drawbox/src/commonMain/kotlin/io/github/markyav/drawbox/controller/DrawBoxConnectionState.kt
@@ -1,6 +1,0 @@
-package io.github.markyav.drawbox.controller
-
-sealed interface DrawBoxConnectionState {
-    object Disconnected : DrawBoxConnectionState
-    data class Connected(val size: Int, val alpha: Float = 1f) : DrawBoxConnectionState // it is square
-}
