--- drawbox/src/commonMain/kotlin/io/github/markyav/drawbox/controller/DrawBoxSubscription.kt
@@ -1,6 +1,0 @@
-package io.github.markyav.drawbox.controller
-
-sealed interface DrawBoxSubscription {
-    object DynamicUpdate : DrawBoxSubscription
-    object FinishDrawingUpdate : DrawBoxSubscription
-}
