--- drawbox/src/commonMain/kotlin/io/github/markyav/drawbox/model/CanvasAction.kt
@@ -1,6 +1,0 @@
-package io.github.markyav.drawbox.model
-
-sealed class CanvasAction {
-    data class Draw(val path: PathWrapper) : CanvasAction()
-    data class Erase(val erased: List<PathWrapper>) : CanvasAction()
-}
