--- drawbox/src/commonMain/kotlin/io/github/markyav/drawbox/model/CanvasTool.kt
@@ -1,6 +1,0 @@
-package io.github.markyav.drawbox.model
-
-enum class CanvasTool {
-    BRUSH,
-    ERASER;
-}
