--- drawbox/src/commonMain/kotlin/io/github/markyav/drawbox/model/PathWrapper.kt
@@ -1,12 +1,0 @@
-package io.github.markyav.drawbox.model
-
-import androidx.compose.runtime.snapshots.SnapshotStateList
-import androidx.compose.ui.geometry.Offset
-import androidx.compose.ui.graphics.Color
-
-data class PathWrapper(
-    var points: List<Offset>,
-    val strokeWidth: Float = 5f,
-    val strokeColor: Color,
-    val alpha: Float = 1f
-)
