--- drawbox/src/commonMain/kotlin/io/github/markyav/drawbox/util/Util.kt
@@ -1,36 +1,0 @@
-package io.github.markyav.drawbox.util
-
-import androidx.compose.ui.geometry.Offset
-import androidx.compose.ui.graphics.Path
-
-fun createPath(points: List<Offset>): Path {
-    return Path().apply {
-        if (points.size > 1) {
-            var oldPoint: Offset? = null
-            this.moveTo(points[0].x, points[0].y)
-            for (i in 1 until points.size) {
-                val point: Offset = points[i]
-                oldPoint?.let {
-                    val midPoint = calculateMidpoint(it, point)
-                    if (i == 1) {
-                        this.lineTo(midPoint.x, midPoint.y)
-                    } else {
-                        this.quadraticBezierTo(it.x, it.y, midPoint.x, midPoint.y)
-                    }
-                }
-                oldPoint = point
-            }
-            oldPoint?.let { this.lineTo(it.x, oldPoint.y) }
-        }
-    }
-}
-
-private fun calculateMidpoint(start: Offset, end: Offset) =
-    Offset((start.x + end.x) / 2, (start.y + end.y) / 2)
-
-fun <E> MutableList<E>.addNotNull(element: E?): Boolean {
-    return if (element == null)
-        false
-    else
-        this.add(element)
-}
