+++ drawbox/src/commonMain/kotlin/uk/codecymru/drawbox/controller/DrawBoxConnectionState.kt
@@ -1,0 +1,6 @@
+package uk.codecymru.drawbox.controller
+
+sealed interface DrawBoxConnectionState {
+    object Disconnected : DrawBoxConnectionState
+    data class Connected(val size: Int, val alpha: Float = 1f) : DrawBoxConnectionState // it is square
+}
