+++ drawbox/src/commonMain/kotlin/uk/codecymru/drawbox/controller/DrawBoxSubscription.kt
@@ -1,0 +1,6 @@
+package uk.codecymru.drawbox.controller
+
+sealed interface DrawBoxSubscription {
+    object DynamicUpdate : DrawBoxSubscription
+    object FinishDrawingUpdate : DrawBoxSubscription
+}
