+++ drawbox/src/commonMain/kotlin/uk/codecymru/drawbox/model/CanvasAction.kt
@@ -1,0 +1,6 @@
+package uk.codecymru.drawbox.model
+
+sealed class CanvasAction {
+    data class Draw(val path: PathWrapper) : CanvasAction()
+    data class Erase(val erased: List<PathWrapper>) : CanvasAction()
+}
