+++ drawbox/src/commonMain/kotlin/uk/codecymru/drawbox/model/CanvasTool.kt
@@ -1,0 +1,6 @@
+package uk.codecymru.drawbox.model
+
+enum class CanvasTool {
+    BRUSH,
+    ERASER;
+}
