--- sample/android/src/main/java/io/github/markyav/drawbox/android/MainActivity.kt
@@ -1,18 +1,0 @@
-package io.github.markyav.drawbox.android
-
-import android.os.Bundle
-import androidx.activity.compose.setContent
-import androidx.appcompat.app.AppCompatActivity
-import androidx.compose.material.MaterialTheme
-import io.github.markyav.drawbox.android.app.MainScreen
-
-class MainActivity : AppCompatActivity() {
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContent {
-            MaterialTheme {
-                MainScreen()
-            }
-        }
-    }
-}
