+++ sample/android/src/main/java/uk/codecymru/drawbox/android/MainActivity.kt
@@ -1,0 +1,18 @@
+package uk.codecymru.drawbox.android
+
+import android.os.Bundle
+import androidx.activity.compose.setContent
+import androidx.appcompat.app.AppCompatActivity
+import androidx.compose.material.MaterialTheme
+import uk.codecymru.drawbox.android.app.MainScreen
+
+class MainActivity : AppCompatActivity() {
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContent {
+            MaterialTheme {
+                MainScreen()
+            }
+        }
+    }
+}
